traffic_sign/requirements.txt
streamlit
tensorflow
opencv-python-headless
numpy==1.19.2
Pillow
validators
requests